backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
faiss-cpu==1.8.0
sentence-transformers==2.5.1
python-dotenv==1.0.1
python-multipart==0.0.9
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
sqlalchemy==2.0.29
pypdf==5.3.0
